+++ docs/changelog.rst
@@ -9,11 +9,34 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 Unreleased
 ==========
 
+New
+---
+- Further automate release mechanism. (#1019)
+
+Changed
+-------
+
+Fixed
+-----
+
+Deprecated
+----------
+
+v0.10.0 (2024-06-14)
+====================
+
+New
+---
+- New `PredefinedCatalog` class to handle predefined catalog version based on pooch registry files. (#849)
+
+
 Changed
 -------
 - Development environment is now set up via pixi instead of mamba / conda. See the documentation for more information on how to install.
 - Use the native data CRS when determining zoom levels over the data catalog crs. (#851)
 - Improved `flw.d8_from_dem` method with different options to use river vector data to aid the flow direction derivation. (#305)
+- DataCatalog.predefined_catalogs retrieves predefined_catalogs specified in predefined_catalogs.py. There is no need for setting the predefined_catalogs anymore. (#844)
+
 
 Fixed
 -----
@@ -24,6 +47,11 @@ Fixed
 - Allow for string format in zoom_level path, e.g. `{zoom_level:02d}` (#851)
 - Fixed incorrect renaming of single variable raster datasets (#883)
 - Provide better error message for 0D geometry arrays in GeoDataset (#885)
+- Fixed index error when the number of peaks varies between stations in get_hydrographs method (#933)
+
+Deprecated
+----------
+- The `DataCatalog.from_archive` method is deprecated. Use `DataCatalog.from_yml` with the root pointing to the archive instead. (#849)
 
 v0.9.4 (2024-02-26)
 ===================